Modify hasIccCard method definition.
Modify hasIccCard() definition to change slotId
type from long to int.

+++ telephony/java/android/telephony/TelephonyManager.java
@@ -1400,7 +1400,7 @@ public class TelephonyManager {
      */
     /** {@hide} */
     // FIXME Input argument slotId should be of type int
-    public boolean hasIccCard(long slotId) {
+    public boolean hasIccCard(int slotId) {
 
         try {
             return getITelephony().hasIccCardUsingSlotId(slotId);
